Move all cluster utils tests into the same file
