+++ src/utils/__tests__/clusterUtils.ts
@@ -1,3 +1,4 @@
+import { IState } from 'reducers/types';
 import { Providers } from 'shared/constants';
 import { INodePool, PropertiesOf } from 'shared/types';
 import {
@@ -5,7 +6,10 @@ import {
   v4AWSClusterStatusResponse,
   v5ClusterResponse,
 } from 'testUtils/mockHttpCalls';
+import preloginState from 'testUtils/preloginState';
 import {
+  canClusterUpgrade,
+  computeCapabilities,
   getClusterLatestCondition,
   getInstanceTypesForProvider,
   guessProviderFromNodePools,
@@ -15,6 +19,155 @@ import {
 } from 'utils/clusterUtils';
 
 describe('clusterUtils', () => {
+  describe('canClusterUpgrade', () => {
+    describe('on azure', () => {
+      it('is true for any version', () => {
+        const can = canClusterUpgrade('8.1.0', '9.0.0', 'azure');
+        expect(can).toBe(true);
+      });
+    });
+
+    describe('on aws', () => {
+      it('is true when going from <10.0.0 to <10.0.0', () => {
+        const can = canClusterUpgrade('8.1.0', '9.0.0', 'aws');
+        expect(can).toBe(true);
+      });
+
+      it('is true when going from >10.0.0 to >10.0.0', () => {
+        const can = canClusterUpgrade('10.1.0', '11.0.0', 'aws');
+        expect(can).toBe(true);
+      });
+
+      it('is false when going from <10.0.0 to >10.0.0', () => {
+        const can = canClusterUpgrade('9.1.0', '11.0.0', 'aws');
+        expect(can).toBe(false);
+      });
+    });
+
+    describe('on kvm', () => {
+      it('is true for any version', () => {
+        const can = canClusterUpgrade('8.0.0', '11.0.0', 'kvm');
+        expect(can).toBe(true);
+      });
+    });
+  });
+
+  describe('computeCapabilities', () => {
+    describe('hasOptionalIngress', () => {
+      describe('on azure', () => {
+        it('is false for Azure below 12.0.0', () => {
+          const capabilities = computeCapabilities(
+            getEmptyStateWithProvider('azure')
+          )('11.0.0', 'azure');
+          expect(capabilities.hasOptionalIngress).toBe(false);
+        });
+
+        it('is true for Azure at 12.0.0', () => {
+          const capabilities = computeCapabilities(
+            getEmptyStateWithProvider('azure')
+          )('12.0.0', 'azure');
+          expect(capabilities.hasOptionalIngress).toBe(true);
+        });
+
+        it('is true for Azure above 12.0.0', () => {
+          const capabilities = computeCapabilities(
+            getEmptyStateWithProvider('azure')
+          )('13.0.0', 'azure');
+          expect(capabilities.hasOptionalIngress).toBe(true);
+        });
+      });
+
+      describe('on aws', () => {
+        it('is false for AWS below 10.1.0', () => {
+          const capabilities = computeCapabilities(
+            getEmptyStateWithProvider('aws')
+          )('9.0.0', 'aws');
+          expect(capabilities.hasOptionalIngress).toBe(false);
+        });
+
+        it('is true for AWS at 10.1.0', () => {
+          const capabilities = computeCapabilities(
+            getEmptyStateWithProvider('aws')
+          )('10.1.0', 'aws');
+          expect(capabilities.hasOptionalIngress).toBe(true);
+        });
+
+        it('is true for AWS above 10.1.0', () => {
+          const capabilities = computeCapabilities(
+            getEmptyStateWithProvider('aws')
+          )('11.1.0', 'aws');
+          expect(capabilities.hasOptionalIngress).toBe(true);
+        });
+      });
+
+      describe('on kvm', () => {
+        it('is false for KVM below 12.2.0', () => {
+          const capabilities = computeCapabilities(
+            getEmptyStateWithProvider('kvm')
+          )('11.0.0', 'kvm');
+          expect(capabilities.hasOptionalIngress).toBe(false);
+        });
+
+        it('is true for KVM at 12.2.0', () => {
+          const capabilities = computeCapabilities(
+            getEmptyStateWithProvider('kvm')
+          )('12.2.0', 'kvm');
+          expect(capabilities.hasOptionalIngress).toBe(true);
+        });
+
+        it('is true for KVM above 12.2.0', () => {
+          const capabilities = computeCapabilities(
+            getEmptyStateWithProvider('kvm')
+          )('13.0.0', 'kvm');
+          expect(capabilities.hasOptionalIngress).toBe(true);
+        });
+      });
+    });
+
+    describe('supportsHAMasters', () => {
+      describe('on azure', () => {
+        it('is false for Azure at any version', () => {
+          const capabilities = computeCapabilities(
+            getEmptyStateWithProvider('azure')
+          )('8.1.0', 'azure');
+          expect(capabilities.supportsHAMasters).toBe(false);
+        });
+      });
+
+      describe('on aws', () => {
+        it('is false for AWS below 9.0.0', () => {
+          const capabilities = computeCapabilities(
+            getEmptyStateWithProvider('aws')
+          )('9.0.0', 'aws');
+          expect(capabilities.supportsHAMasters).toBe(false);
+        });
+
+        it('is true for AWS at 11.4.0', () => {
+          const capabilities = computeCapabilities(
+            getEmptyStateWithProvider('aws')
+          )('11.4.0', 'aws');
+          expect(capabilities.supportsHAMasters).toBe(true);
+        });
+
+        it('is true for AWS above 13.0.0', () => {
+          const capabilities = computeCapabilities(
+            getEmptyStateWithProvider('aws')
+          )('13.0.0', 'aws');
+          expect(capabilities.supportsHAMasters).toBe(true);
+        });
+      });
+
+      describe('on kvm', () => {
+        it('is false for KVM at any version', () => {
+          const capabilities = computeCapabilities(
+            getEmptyStateWithProvider('kvm')
+          )('8.0.0', 'kvm');
+          expect(capabilities.supportsHAMasters).toBe(false);
+        });
+      });
+    });
+  });
+
   describe('getClusterLatestCondition', () => {
     it('gets the latest cluster condition, on a v5 cluster', () => {
       const cluster = (Object.assign(
@@ -258,3 +411,21 @@ describe('clusterUtils', () => {
     });
   });
 });
+
+function getEmptyStateWithProvider(
+  provider: PropertiesOf<typeof Providers>
+): IState {
+  return {
+    ...preloginState,
+    main: {
+      ...preloginState.main,
+      info: {
+        ...preloginState.main.info,
+        general: {
+          ...preloginState.main.info.general,
+          provider,
+        },
+      },
+    },
+  };
+}
